Screenshot script
Made a script that can take and delete screenshots. the screenshots gets saved to a temporaryCachePath so they also get deleted on application reset and close. there is also function for manual deletion of files for start of a game.

Made a few folders for scripts in the assets menu