Finished the score calculation and taking screenshots
Made 2 unity events, one for taking screenshots and the other for deleting all data. Both scripts are static classed so I had to create a hook that subscirbes the proper functions to the corresponding event.

created some logic for getting points. The posePoint class is put on a dino with an animator. In the animator per animation you can create a curve with Pose as the parameter and during the animation it changes the value of posepoints based on the curve.